refactor(cleanup): rename Algorithms to Context
