+++ Turner.Infrastructure.Crud/Context/IEntityContext.cs
@@ -5,7 +5,7 @@ using System.Linq.Expressions;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Turner.Infrastructure.Crud.Algorithms
+namespace Turner.Infrastructure.Crud.Context
 {
     public interface IEntityContext
     {
