+++ Turner.Infrastructure.Crud/CrudConfiguration.cs
@@ -2,7 +2,7 @@
 using System;
 using System.Linq;
 using System.Reflection;
-using Turner.Infrastructure.Crud.Algorithms;
+using Turner.Infrastructure.Crud.Context;
 using Turner.Infrastructure.Crud.Configuration;
 using Turner.Infrastructure.Crud.Errors;
 using Turner.Infrastructure.Crud.Requests;
