+++ Turner.Infrastructure.Crud/Extensions/SelectorExtensions.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using Turner.Infrastructure.Crud.Algorithms;
+using Turner.Infrastructure.Crud.Context;
 using Turner.Infrastructure.Crud.Configuration;
 
 namespace Turner.Infrastructure.Crud
