+++ Turner.Infrastructure.Crud/Requests/DeleteAllRequestHandler.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using Turner.Infrastructure.Crud.Algorithms;
+using Turner.Infrastructure.Crud.Context;
 using Turner.Infrastructure.Crud.Configuration;
 using Turner.Infrastructure.Mediator;
