+++ Turner.Infrastructure.Crud/Requests/GetRequestHandler.cs
@@ -2,7 +2,7 @@
 using AutoMapper.QueryableExtensions;
 using System.Linq;
 using System.Threading.Tasks;
-using Turner.Infrastructure.Crud.Algorithms;
+using Turner.Infrastructure.Crud.Context;
 using Turner.Infrastructure.Crud.Configuration;
 using Turner.Infrastructure.Crud.Errors;
 using Turner.Infrastructure.Crud.Exceptions;
