+++ Turner.Infrastructure.Crud/Requests/PagedFindRequestHandler.cs
@@ -1,7 +1,7 @@
 ﻿using AutoMapper;
 using System.Linq;
 using System.Threading.Tasks;
-using Turner.Infrastructure.Crud.Algorithms;
+using Turner.Infrastructure.Crud.Context;
 using Turner.Infrastructure.Crud.Configuration;
 using Turner.Infrastructure.Crud.Errors;
 using Turner.Infrastructure.Crud.Exceptions;
