+++ Turner.Infrastructure.Crud/Requests/SynchronizeRequestHandler.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
 using System.Threading.Tasks;
-using Turner.Infrastructure.Crud.Algorithms;
+using Turner.Infrastructure.Crud.Context;
 using Turner.Infrastructure.Crud.Configuration;
 using Turner.Infrastructure.Mediator;
